Use a "system" user for running sudo
The old method of always doing usermod could have issues where sudo
would refuse to run items when a user didn't exist in /etc/passwd.

So now a user with uid < 100 is created which is intended to create the
new user.

This also fixes the bug where we started the container as root.

Signed-off-by: Randy Witt <randy.e.witt@linux.intel.com>