usersetup.py: Error if workdir has non-zero uid:gid
When the uid:gid is determined by looking at the ownership of the
workdir, error if the uid:gid is non-zero. This is because the user
running bitbake must be non-zero.

This is also hopefully a more helpful error message for when the user
mistypes the volume to docker using "-v" or doesn't specify an absolute
path.

Signed-off-by: Randy Witt <randy.e.witt@intel.com>